+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added ISS decoding for HVC and SVC exceptions.
 - Added ISS decoding for MRS and MSR exceptions, including system register names.
 - Added ISS decoding for Pointer Authentication failures.
+- Added ISS decoding for floating-point exceptions.
 
 ## 0.1.0